+++ tfjs-core/src/ops/ops.ts
@@ -18,6 +18,8 @@
 // Modularized ops.
 export {add} from './add';
 export {addN} from './add_n';
+export {avgPool} from './avg_pool';
+export {avgPool3d} from './avg_pool_3d';
 export {batchToSpaceND} from './batch_to_space_nd';
 export {batchNorm} from './batchnorm';
 export {batchNorm2d} from './batchnorm2d';
@@ -50,6 +52,9 @@ export {lessEqual} from './less_equal';
 export {localResponseNormalization} from './local_response_normalization';
 export {matMul} from './mat_mul';
 export {max} from './max';
+export {maxPool} from './max_pool';
+export {maxPool3d} from './max_pool_3d';
+export {maxPoolWithArgmax} from './max_pool_with_argmax';
 export {multinomial} from './multinomial';
 export {notEqual} from './not_equal';
 export {oneHot} from './one_hot';
@@ -59,6 +64,7 @@ export {pad1d} from './pad1d';
 export {pad2d} from './pad2d';
 export {pad3d} from './pad3d';
 export {pad4d} from './pad4d';
+export {pool} from './pool';
 export {rand} from './rand';
 export {randomGamma} from './random_gamma';
 export {randomNormal} from './random_normal';
@@ -75,7 +81,6 @@ export {truncatedNormal} from './truncated_normal';
 export * from './boolean_mask';
 export * from './complex_ops';
 export * from './reverse';
-export * from './pool';
 export * from './slice';
 export * from './unary_ops';
 export * from './reduction_ops';
