+++ tfjs-core/src/public/chained_ops/register_all_chained_ops_test.ts
@@ -25,6 +25,7 @@ import {ALL_ENVS, describeWithFlags} from '../../jasmine_util';
 
 const CHAINED_OPS = [
   'add',
+  'avgPool',
   'batchNorm',
   'batchToSpaceND',
   'broadcastTo',
@@ -45,10 +46,12 @@ const CHAINED_OPS = [
   'lessEqual',
   'localResponseNormalization',
   'matMul',
+  'max',
+  'maxPool',
   'notEqual',
   'oneHot',
   'pad',
-  'max',
+  'pool',
   'separableConv2d',
   'spaceToBatchND',
   'split',
