+++ tfjs-core/src/register_all_gradients.ts
@@ -16,6 +16,8 @@
  */
 import {addGradConfig} from './gradients/Add_grad';
 import {addNGradConfig} from './gradients/AddN_grad';
+import {avgPool3DGradConfig} from './gradients/AvgPool3D_grad';
+import {avgPoolGradConfig} from './gradients/AvgPool_grad';
 import {batchMatMulGradConfig} from './gradients/BatchMatMul_grad';
 import {batchToSpaceNDGradConfig} from './gradients/BatchToSpaceND_grad';
 import {broadcastToGradConfig} from './gradients/BroadcastTo_grad';
@@ -30,6 +32,8 @@ import {greaterEqualGradConfig} from './gradients/GreaterEqual_grad';
 import {identityGradConfig} from './gradients/Identity_grad';
 import {lrnGradConfig} from './gradients/LRN_grad';
 import {maxGradConfig} from './gradients/Max_grad';
+import {maxPool3DGradConfig} from './gradients/MaxPool3D_grad';
+import {maxPoolGradConfig} from './gradients/MaxPool_grad';
 import {oneHotGradConfig} from './gradients/OneHot_grad';
 import {padV2GradConfig} from './gradients/PadV2_grad';
 import {spaceToBatchNDGradConfig} from './gradients/SpaceToBatchND_grad';
@@ -44,18 +48,34 @@ import {registerGradient} from './kernel_registry';
 
 // Export all kernel configs here so that the package can auto register them
 const gradConfigs: GradConfig[] = [
-  addGradConfig,          addNGradConfig,
-  batchMatMulGradConfig,  batchToSpaceNDGradConfig,
-  broadcastToGradConfig,  concatGradConfig,
-  conv2DGradConfig,       conv2DBackpropInputGradConfig,
-  conv3DGradConfig,       depthwiseConv2dNativeGradConfig,
-  divGradConfig,          fusedBatchNormGradConfig,
-  greaterEqualGradConfig, identityGradConfig,
-  lrnGradConfig,          oneHotGradConfig,
-  padV2GradConfig,        splitVGradConfig,
-  maxGradConfig,          spaceToBatchNDGradConfig,
-  squareGradConfig,       squaredDifferenceGradConfig,
-  tileGradConfig,         transposeGradConfig,
+  addGradConfig,
+  addNGradConfig,
+  avgPoolGradConfig,
+  avgPool3DGradConfig,
+  batchMatMulGradConfig,
+  batchToSpaceNDGradConfig,
+  broadcastToGradConfig,
+  concatGradConfig,
+  conv2DGradConfig,
+  conv2DBackpropInputGradConfig,
+  conv3DGradConfig,
+  depthwiseConv2dNativeGradConfig,
+  divGradConfig,
+  fusedBatchNormGradConfig,
+  greaterEqualGradConfig,
+  identityGradConfig,
+  lrnGradConfig,
+  maxGradConfig,
+  maxPoolGradConfig,
+  maxPool3DGradConfig,
+  oneHotGradConfig,
+  padV2GradConfig,
+  spaceToBatchNDGradConfig,
+  splitVGradConfig,
+  squareGradConfig,
+  squaredDifferenceGradConfig,
+  tileGradConfig,
+  transposeGradConfig,
   subGradConfig
 ];
