+++ tfjs-core/src/tests.ts
@@ -42,6 +42,8 @@ import './ops/add_n_test';
 import './ops/add_test';
 import './ops/arithmetic_test';
 import './ops/array_ops_test';
+import './ops/avg_pool_3d_test';
+import './ops/avg_pool_test';
 import './ops/axis_util_test';
 import './ops/batch_to_space_nd_test';
 import './ops/batchnorm_test';
@@ -83,6 +85,9 @@ import './ops/logical_ops_test';
 import './ops/loss_ops_test';
 import './ops/lstm_test';
 import './ops/mat_mul_test';
+import './ops/max_pool_3d_test';
+import './ops/max_pool_test';
+import './ops/max_pool_with_argmax_test';
 import './ops/moving_average_test';
 import './ops/multinomial_test';
 import './ops/not_equal_test';
